Allow setting resizeToAvoidBottomInset
The default setting of false causes issues on mobile, since the default header image height causes the password field to be hidden on smaller phones (including the pixel 4a or other lower end devices). This causes a poor user experience.

This tweak allows passing through the resizeToAvoidBottomInset externally, with the current behaviour as default to avoid breaking changes.